News/category modul

+++ database/migrations/2023_04_01_182043_create_category_translations_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('category_translations', function (Blueprint $table) {
             $table->id();
 
-            $table->integer('category_id')->unsigned();
+            $table->unsignedBigInteger('category_id')->unsigned();
             $table->string('locale')->index();
 
             $table->string('name')->nullable()->default(NULL);
